Add files via upload
A. How many seconds are there in 42 minutes 42 seconds?

B. The volume of a sphere with radius r is (4/3)πr. What is the volume of a sphere with radius 4? With radius 6? (Hint: pi=3.14159265 but import and use the exact number from the necessary module)

C. Show in Python for -40° is the same temperature in both Fahrenheit and Celsius (Run the conversion for -40°).

D. Suppose that you have a rectangular prism R with sides a, 2a, 3a and a cube C with side a. How many of C can fit in R?